requirements.txt
Django==4.2.3
Pillow==10.0.0
environs==9.5.0
django-phonenumber-field==7.1.0